+++ crates/risc0/src/host/mod.rs
@@ -1,14 +1,19 @@
 //! This module implements the [`ZkvmHost`] trait for the RISC0 VM.
 
 mod bonsai;
+mod boundless;
+/// Configuration types for RISC0 provers
+pub mod config;
 mod local;
+mod pricing_service;
 
-use std::path::Path;
 use std::time::{SystemTime, UNIX_EPOCH};
-use std::{env, fs, mem};
+use std::{fs, mem};
 
 use bonsai::BonsaiProver;
 use borsh::BorshDeserialize;
+use boundless::BoundlessProver;
+use config::Risc0ProverConfig;
 use local::LocalProver;
 use risc0_zkvm::sha::Digest;
 use risc0_zkvm::{AssumptionReceipt, VerifierContext};
@@ -16,7 +21,7 @@ use sov_db::ledger_db::LedgerDB;
 use sov_rollup_interface::zk::{Proof, ProofWithJob, ReceiptType, Zkvm, ZkvmHost};
 use sov_rollup_interface::Network;
 use tokio::sync::oneshot;
-use tracing::{debug, info};
+use tracing::info;
 use uuid::Uuid;
 
 use crate::guest::Risc0Guest;
@@ -28,20 +33,25 @@ pub struct Risc0Host {
     env: Vec<u8>,
     assumptions: Vec<AssumptionReceipt>,
     prover: Prover,
+    tx_backup_dir: Option<std::path::PathBuf>,
 }
 
 impl Risc0Host {
     /// Create a new Risc0Host to prove the given binary.
-    pub fn new(ledger_db: LedgerDB, network: Network) -> Self {
-        let prover = match std::env::var("RISC0_PROVER") {
-            Ok(prover) => match prover.as_str() {
-                "bonsai" => Prover::Bonsai(BonsaiProver::new(ledger_db)),
-                "ipc" => Prover::Local(LocalProver::new(network)),
-                _ => panic!("Invalid prover specified: {}", prover),
-            },
-            Err(_) => {
-                debug!("No prover specified.");
-                Prover::Local(LocalProver::new(network))
+    pub async fn new(
+        ledger_db: LedgerDB,
+        network: Network,
+        config: config::Risc0HostConfig,
+    ) -> Self {
+        let prover = match config.prover {
+            Risc0ProverConfig::Boundless(boundless_config) => {
+                Prover::Boundless(BoundlessProver::new(ledger_db, *boundless_config).await)
+            }
+            Risc0ProverConfig::Bonsai(bonsai_config) => {
+                Prover::Bonsai(BonsaiProver::new(ledger_db, bonsai_config))
+            }
+            Risc0ProverConfig::Local(local_config) => {
+                Prover::Local(LocalProver::new(network, local_config))
             }
         };
 
@@ -49,10 +59,12 @@ impl Risc0Host {
             env: Default::default(),
             assumptions: vec![],
             prover,
+            tx_backup_dir: config.tx_backup_dir,
         }
     }
 }
 
+#[async_trait::async_trait]
 impl ZkvmHost for Risc0Host {
     type Guest = Risc0Guest;
 
@@ -72,7 +84,7 @@ impl ZkvmHost for Risc0Host {
         self.assumptions.push(receipt.into());
     }
 
-    fn run(
+    async fn run(
         &mut self,
         job_id: Uuid,
         elf: Vec<u8>,
@@ -82,8 +94,8 @@ impl ZkvmHost for Risc0Host {
         let input = mem::take(&mut self.env);
         let assumptions = mem::take(&mut self.assumptions);
 
-        if let Ok(backup_dir) = env::var("TX_BACKUP_DIR") {
-            let input_path = Path::new(&backup_dir).join(format!(
+        if let Some(backup_dir) = &self.tx_backup_dir {
+            let input_path = backup_dir.join(format!(
                 "{}-proof-input.bin",
                 SystemTime::now()
                     .duration_since(UNIX_EPOCH)
@@ -104,6 +116,15 @@ impl ZkvmHost for Risc0Host {
                 );
                 bonsai.prove(job_id, elf, input, assumptions, receipt_type)
             }
+            Prover::Boundless(boundless) => {
+                assert!(
+                    with_prove,
+                    "Boundless prover must always be run with prove set to true"
+                );
+                boundless
+                    .prove(job_id, elf, input, assumptions, receipt_type)
+                    .await
+            }
         }
     }
 
@@ -117,11 +138,7 @@ impl ZkvmHost for Risc0Host {
     fn start_session_recovery(
         &self,
     ) -> Result<Vec<oneshot::Receiver<ProofWithJob>>, anyhow::Error> {
-        let Prover::Bonsai(prover) = &self.prover else {
-            info!("Skipping proving recovery...");
-            return Ok(vec![]);
-        };
-        prover.start_recovery()
+        self.prover.start_prover_session_recovery()
     }
 }
 
@@ -171,9 +188,28 @@ impl Zkvm for Risc0Host {
 
 /// Supported `Prover` types
 #[derive(Clone)]
+#[allow(clippy::large_enum_variant)]
 pub enum Prover {
     /// Local prover
     Local(LocalProver),
     /// Bonsai prover
     Bonsai(BonsaiProver),
+    /// Boundless prover network
+    Boundless(BoundlessProver),
+}
+
+impl Prover {
+    /// Start recovery for prover if it supports it
+    pub fn start_prover_session_recovery(
+        &self,
+    ) -> anyhow::Result<Vec<oneshot::Receiver<ProofWithJob>>> {
+        match self {
+            Prover::Local(_) => {
+                info!("Skipping proving recovery...");
+                Ok(vec![])
+            }
+            Prover::Boundless(prover) => prover.start_recovery(),
+            Prover::Bonsai(prover) => prover.start_recovery(),
+        }
+    }
 }