+++ crates/sovereign-sdk/full-node/db/sov-db/src/schema/tables.rs
@@ -23,8 +23,8 @@ use super::types::batch_proof::{StoredBatchProof, StoredVerifiedProof};
 use super::types::l2_block::StoredL2Block;
 use super::types::light_client_proof::StoredLightClientProof;
 use super::types::{
-    AccessoryKey, AccessoryStateValue, BonsaiSession, DbHash, JmtValue, L1Height, L2BlockNumber,
-    L2HeightAndIndex, L2HeightRange, L2HeightStatus, SlotNumber, StateKey,
+    AccessoryKey, AccessoryStateValue, BonsaiSession, BoundlessSession, DbHash, JmtValue, L1Height,
+    L2BlockNumber, L2HeightAndIndex, L2HeightRange, L2HeightStatus, SlotNumber, StateKey,
 };
 
 /// A list of all tables used by the StateDB. These tables store rollup state - meaning
@@ -101,6 +101,7 @@ pub const BATCH_PROVER_LEDGER_TABLES: &[&str] = &[
     LastPrunedBlock::table_name(),
     PendingBonsaiSessionByJobId::table_name(),
     PendingL1SubmissionJobs::table_name(),
+    PendingBoundlessSessionByJobId::table_name(),
     ProofByJobId::table_name(),
     ProverLastScannedSlot::table_name(),
     ProverPendingCommitments::table_name(),
@@ -121,11 +122,12 @@ pub const BATCH_PROVER_LEDGER_TABLES: &[&str] = &[
 pub const LIGHT_CLIENT_PROVER_LEDGER_TABLES: &[&str] = &[
     // Don't know if this will be needed
     CommitmentMerkleRoots::table_name(),
+    PendingBonsaiSessionByJobId::table_name(),
+    PendingBoundlessSessionByJobId::table_name(),
     ExecutedMigrations::table_name(),
     LightClientProofBySlotNumber::table_name(),
     ProverLastScannedSlot::table_name(),
     SlotByHash::table_name(),
-    PendingBonsaiSessionByJobId::table_name(),
     // #### TESTS RELATED TABLES ####
     #[cfg(test)]
     TestTableOld::table_name(),
@@ -163,6 +165,7 @@ pub const LEDGER_TABLES: &[&str] = &[
     ProofsBySlotNumberV2::table_name(),
     ProverLastScannedSlot::table_name(),
     ProverPendingCommitments::table_name(),
+    PendingBoundlessSessionByJobId::table_name(),
     ProverStateDiffs::table_name(),
     SequencerCommitmentByIndex::table_name(),
     ShortHeaderProofBySlotHash::table_name(),
@@ -384,9 +387,14 @@ define_table_with_default_codec!(
     (PendingL1SubmissionJobs) Uuid => ()
 );
 
-define_table_with_default_codec!(
+define_table_with_seek_key_codec!(
     /// Pending Bonsai proving sessions by job id
     (PendingBonsaiSessionByJobId) Uuid => BonsaiSession
+);
+
+define_table_with_seek_key_codec!(
+    /// Pending Boundless proving sessions by job id
+    (PendingBoundlessSessionByJobId) Uuid => BoundlessSession
 );
 
 define_table_with_default_codec!(
@@ -509,7 +517,7 @@ impl KeyEncoder<JmtNodes> for NodeKey {
     fn encode_key(&self) -> sov_schema_db::schema::Result<Vec<u8>> {
         // 8 bytes for version, 4 each for the num_nibbles and bytes.len() fields, plus 1 byte per byte of nibblepath
         let mut output =
-            Vec::with_capacity(8 + 4 + 4 + ((self.nibble_path().num_nibbles() + 1) / 2));
+            Vec::with_capacity(8 + 4 + 4 + self.nibble_path().num_nibbles().div_ceil(2));
         let version = self.version().to_be_bytes();
         output.extend_from_slice(&version);
         BorshSerialize::serialize(self.nibble_path(), &mut output)?;